requirements.txt

reflex==0.8.14
openai
anthropic
google-generativeai
requests
httpx
mcp